backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pyjwt
passlib[bcrypt]
requests
python-dotenv
pandas
